refactor business logic into a controller class
