+++ src/server/server.ts
@@ -7,35 +7,20 @@ import * as primusEmit from "primus-emit"
 import * as primusSparkLatency from "primus-spark-latency"
 // @ts-ignore
 import _ from "lodash"
-import { Keccak } from 'sha3'
-// @ts-ignore
-import { ec as EC } from "elliptic"
-// @ts-ignore
-import { KeyPair } from "elliptic/lib/elliptic/ec"
+import Controller from "./Controller";
 import { createServer } from "http"
 import routes from "./routes"
-import Collection from './collection'
 import {
   banned,
   reserved,
   trusted
 } from "./utils/config"
-import Node from "./node"
-import { BasicStatsResponse } from "./interfaces/BasicStatsResponse";
 import { Proof } from "./interfaces/Proof";
 import { NodeResponseLatency } from "./interfaces/NodeResponseLatency";
-import { Pending } from "./interfaces/Pending";
-import { NodeInfo } from "./interfaces/NodeInfo";
-import { ChartData } from "./interfaces/ChartData";
-import { BlockStats } from "./interfaces/BlockStats";
 import { StatsWrapped } from "./interfaces/StatsWrapped";
-import { NodePong } from "./interfaces/NodePong";
 import { ClientPong } from "./interfaces/ClientPong";
 import { NodeResponsePing } from "./interfaces/NodeResponsePing";
-import { NodeStats } from "./interfaces/NodeStats";
 import { NodeResponseStats } from "./interfaces/NodeResponseStats";
-import { NodeInformation } from "./interfaces/NodeInformation";
-import { NodeData } from "./interfaces/NodeData";
 import { Latency } from "./interfaces/Latency";
 import { NodePing } from "./interfaces/NodePing";
 import { NodeResponseBlock } from "./interfaces/NodeResponseBlock";
@@ -44,18 +29,11 @@ import { NodeResponseInfo } from "./interfaces/NodeResponseInfo";
 import { InfoWrapped } from "./interfaces/InfoWrapped";
 import { Sides } from "./statistics/Sides";
 import { Directions } from "./statistics/Directions";
-import { Statistics } from "./statistics/Statistics";
 import { IDictionary } from "./interfaces/IDictionary";
-import { Wrapper } from "./interfaces/Wrapper";
-import { Validators } from "./interfaces/Validators";
-import { Block } from "./interfaces/Block";
-import { Stats } from "./interfaces/Stats";
+import { isInputValid } from "./utils/isInputValid";
 
 // general config
-const clientPingTimeout = 5 * 1000
-const nodeCleanupTimeout = 1000 * 60 * 60
 const port = process.env.PORT || 3000
-const statisticsInterval = 60 * 1000
 
 // add trusted from env
 if (process.env.TRUSTED_ADDRESSES) {
@@ -68,16 +46,12 @@ if (process.env.RESERVED_ADDRESSES) {
   reserved.push(...process.env.RESERVED_ADDRESSES.split(','))
 }
 
-
-
 export default class Server {
 
-  private readonly nodes: Collection
   private readonly api: Primus
   private readonly client: Primus
-  private danglingConnections: IDictionary = {}
-
-  private statistics: Statistics;
+  private readonly danglingConnections: IDictionary = {}
+  private readonly controller: Controller;
 
   constructor() {
     const server = createServer(routes)
@@ -102,9 +76,10 @@ export default class Server {
       pingInterval: false
     })
 
-    this.nodes = new Collection()
-
-    this.statistics = new Statistics(this.nodes);
+    this.controller = new Controller(
+      this.api,
+      this.client
+    )
 
     server.listen(port)
 
@@ -114,131 +89,6 @@ export default class Server {
     )
   }
 
-  static isInputValid(stats: Wrapper): boolean {
-    return (
-      !_.isUndefined(stats) && !_.isUndefined(stats.id)
-    )
-  }
-
-  static isAuthorize(
-    proof: Proof,
-    stats: InfoWrapped
-  ): boolean {
-    let isAuthorized = false
-
-    if (
-      Server.isInputValid(stats) &&
-      !_.isUndefined(proof) &&
-      !_.isUndefined(proof.publicKey) &&
-      !_.isUndefined(proof.signature) &&
-      reserved.indexOf(stats.id) < 0 &&
-      trusted
-        .map(address => address && address.toLowerCase())
-        .indexOf(proof.address) >= 0
-    ) {
-      const hasher = new Keccak(256)
-      hasher.update(JSON.stringify(stats))
-
-      const msgHash = hasher.digest('hex')
-      const secp256k1 = new EC('secp256k1')
-      const pubkeyNoZeroX = proof.publicKey.substr(2)
-
-      let pubkey: KeyPair
-
-      try {
-        pubkey = secp256k1.keyFromPublic(pubkeyNoZeroX, 'hex')
-      } catch (err) {
-        console.error('API', 'SIG', 'Public Key Error', err.message)
-        return false
-      }
-
-      const addressHasher = new Keccak(256)
-      addressHasher.update(pubkeyNoZeroX.substr(2), 'hex')
-
-      const addressHash = addressHasher.digest('hex').substr(24)
-
-      if (!(addressHash.toLowerCase() === proof.address.substr(2).toLowerCase())) {
-        console.error(
-          'API', 'SIG',
-          'Address hash did not match', addressHash,
-          proof.address.substr(2)
-        )
-      }
-
-      const signature: {
-        r: string
-        s: string
-      } = {
-        r: proof.signature.substr(2, 64),
-        s: proof.signature.substr(66, 64)
-      }
-
-      if (!(msgHash === proof.msgHash.substr(2))) {
-        console.error(
-          'API', 'SIG',
-          'Message hash did not match',
-          msgHash, proof.msgHash.substr(2)
-        )
-        return false
-      }
-
-      try {
-        isAuthorized = pubkey.verify(msgHash, signature)
-        if (!isAuthorized) {
-          console.error('API', 'SIG', 'Signature did not verify')
-          return false
-        }
-      } catch (e) {
-        console.error('API', 'SIG', 'Signature Error', e.message)
-        return false
-      }
-    }
-
-    return isAuthorized
-  }
-
-  private clientWrite(payload: object) {
-    this.client.forEach((spark: Primus.spark) => {
-      spark.write(payload)
-
-      this.statistics.add(Sides.Client, Directions.Out)
-    });
-  }
-
-  private wireup(): void {
-    // ping clients
-    setInterval(() => {
-      this.clientWrite({
-        action: 'client-ping',
-        data: {
-          serverTime: _.now()
-        }
-      })
-    }, clientPingTimeout)
-
-    // Cleanup old inactive nodes
-    setInterval(() => {
-      this.clientWrite({
-        action: 'init',
-        data: this.nodes.all()
-      })
-
-      this.nodes.getCharts()
-
-    }, nodeCleanupTimeout)
-
-    // print statistics
-    setInterval(() => {
-      let clients = 0
-      this.client.forEach(() => clients++);
-
-      let nodes = 0;
-      this.api.forEach(() => nodes++);
-
-      this.statistics.print(clients, nodes);
-    }, statisticsInterval)
-  }
-
   private initApi(): void {
     // Init API Socket connection
     this.api.plugin('emit', primusEmit)
@@ -246,9 +96,9 @@ export default class Server {
 
     // Init API Socket events
     this.api.on('connection', (spark: Primus.spark): void => {
-      this.statistics.add(Sides.Node, Directions.In)
+      this.controller.statistics.add(Sides.Node, Directions.In)
 
-      // make connection dangling
+      // make connection dangling by default
       this.danglingConnections[spark.id] = 1
 
       console.success(
@@ -257,7 +107,7 @@ export default class Server {
       )
 
       spark.on('hello', (data: NodeResponseInfo): void => {
-        this.statistics.add(Sides.Node, Directions.In)
+        this.controller.statistics.add(Sides.Node, Directions.In)
 
         const {
           stats,
@@ -269,16 +119,18 @@ export default class Server {
 
         const id = proof.address;
 
-        if (
-          this.handleAuth(id, proof, stats, spark) &&
-          !_.isUndefined(stats.info)
-        ) {
-          this.handleNode(id, proof, stats, spark)
+        const isAuthed = this.controller.handleNodeHello(
+          id, proof, stats, spark
+        )
+
+        if (isAuthed) {
+          // remove dangling connection on success
+          delete (this.danglingConnections[spark.id])
         }
       })
 
       spark.on('block', (data: NodeResponseBlock): void => {
-        this.statistics.add(Sides.Node, Directions.In)
+        this.controller.statistics.add(Sides.Node, Directions.In)
 
         const {
           stats,
@@ -289,15 +141,15 @@ export default class Server {
         } = data
 
         if (
-          Server.isInputValid(stats) &&
+          isInputValid(stats) &&
           !_.isUndefined(stats.block)
         ) {
           const id = proof.address
 
           // handle validator information from block
-          this.handleValidators(stats.block.validators)
+          this.controller.handleNodeBlockValidators(stats.block.validators)
 
-          this.handleBlock(id, spark.address.ip, stats.block)
+          this.controller.handleNodeBlock(id, spark.address.ip, stats.block)
 
         } else {
           console.error('API', 'BLK', 'Block error:', data)
@@ -305,7 +157,7 @@ export default class Server {
       })
 
       spark.on('pending', (data: NodeResponseStats): void => {
-        this.statistics.add(Sides.Node, Directions.In)
+        this.controller.statistics.add(Sides.Node, Directions.In)
 
         const {
           stats,
@@ -316,19 +168,19 @@ export default class Server {
         } = data
 
         if (
-          Server.isInputValid(stats) &&
+          isInputValid(stats) &&
           !_.isUndefined(stats.stats)
         ) {
           const id = proof.address
 
-          this.handlePending(id, stats.stats)
+          this.controller.handleNodePending(id, stats.stats)
         } else {
           console.error('API', 'TXS', 'Pending error:', data)
         }
       })
 
       spark.on('stats', (data: NodeResponseStats): void => {
-        this.statistics.add(Sides.Node, Directions.In)
+        this.controller.statistics.add(Sides.Node, Directions.In)
 
         const {
           stats,
@@ -339,19 +191,19 @@ export default class Server {
         } = data;
 
         if (
-          Server.isInputValid(stats) &&
+          isInputValid(stats) &&
           !_.isUndefined(stats.stats)
         ) {
 
           // why? why not spark.id like everywhere?
           const id = proof.address
 
-          this.handleStatsUpdate(id, stats.stats)
+          this.controller.handleNodeStats(id, stats.stats)
         }
       })
 
       spark.on('node-ping', (data: NodeResponsePing): void => {
-        this.statistics.add(Sides.Node, Directions.In)
+        this.controller.statistics.add(Sides.Node, Directions.In)
 
         const {
           stats,
@@ -361,15 +213,15 @@ export default class Server {
           proof: Proof
         } = data
 
-        if (Server.isInputValid(stats)) {
+        if (isInputValid(stats)) {
           const id = proof.address
 
-          this.handleNodePing(id, stats, spark)
+          this.controller.handleNodePing(id, stats, spark)
         }
       })
 
       spark.on('latency', (data: NodeResponseLatency): void => {
-        this.statistics.add(Sides.Node, Directions.In)
+        this.controller.statistics.add(Sides.Node, Directions.In)
 
         const {
           stats,
@@ -379,21 +231,33 @@ export default class Server {
           proof: Proof
         } = data
 
-        if (Server.isInputValid(stats)) {
+        if (isInputValid(stats)) {
 
           const id = proof.address
 
-          this.handleLatency(id, stats.latency)
+          this.controller.handleNodeLatency(id, stats.latency)
         }
       })
 
       spark.on('end', (): void => {
-        this.statistics.add(Sides.Node, Directions.In)
+        this.controller.statistics.add(Sides.Node, Directions.In)
 
         // use spark id here, we have nothing else
         const id = spark.id;
 
-        this.handleNodeEnd(id, spark.address.ip)
+        // it this was caused by a failed auth do not try this
+        if (this.danglingConnections[id]) {
+          delete (this.danglingConnections[spark.id])
+
+        } else {
+          this.controller.handleNodeEnd(id)
+        }
+
+        console.success(
+          'API', 'CON', 'Node Close:',
+          spark.address.ip, `'${id}'`
+        )
+
       })
     })
   }
@@ -403,7 +267,7 @@ export default class Server {
     this.client.plugin('emit', primusEmit)
 
     this.client.on('connection', (spark: Primus.spark): void => {
-      this.statistics.add(Sides.Client, Directions.In)
+      this.controller.statistics.add(Sides.Client, Directions.In)
 
       console.success(
         'API', 'CON', 'Client Open:',
@@ -411,383 +275,31 @@ export default class Server {
       )
 
       spark.on('ready', (): void => {
-        this.statistics.add(Sides.Client, Directions.In)
+        this.controller.statistics.add(Sides.Client, Directions.In)
 
         const id = spark.id
 
-        this.handleReady(id, spark)
+        this.controller.handleClientReady(id, spark)
       })
 
       spark.on('client-pong', (data: ClientPong): void => {
-        this.statistics.add(Sides.Client, Directions.In)
+        this.controller.statistics.add(Sides.Client, Directions.In)
 
-        this.handleClientPong(data, spark)
+        this.controller.handleClientPong(data, spark)
       })
 
-      spark.on('end', () => {
-
+      spark.on('end', (): void => {
         const id = spark.id;
-        this.handleClientEnd(id, spark.address.ip)
+
+        this.controller.handleClientEnd(id, spark.address.ip)
       })
 
     })
   }
 
-  private initNodes(): void {
-    // Init collections
-    this.nodes.setChartsCallback(
-      (err: Error | string, charts: ChartData): void => {
-        if (err) {
-          console.error('COL', 'CHR', 'Charts error:', err)
-        } else {
-          this.clientWrite({
-            action: 'charts',
-            data: charts
-          })
-        }
-      }
-    )
-  }
-
   public init(): void {
     this.initApi()
     this.initClient()
-    this.initNodes()
-    this.wireup()
-  }
-
-  private handleBlock(
-    id: string,
-    ip: string,
-    block: Block
-  ): void {
-    this.nodes.addBlock(
-      id, block,
-      (err: Error | string, updatedStats: BlockStats) => {
-        if (err) {
-          console.error('API', 'BLK', 'Block error:', err, updatedStats)
-        } else if (updatedStats) {
-
-          this.clientWrite({
-            action: 'block',
-            data: updatedStats
-          })
-
-          console.info('API', 'BLK',
-            'Block:', updatedStats.block['number'],
-            'td:', updatedStats.block['totalDifficulty'],
-            'from:', updatedStats.id, 'ip:', ip)
-
-          this.nodes.getCharts()
-        }
-      },
-      (err: Error | string, highestBlock: number) => {
-        if (err) {
-          console.error(err)
-        } else {
-          this.clientWrite({
-            action: 'lastBlock',
-            number: highestBlock
-          })
-        }
-      }
-    )
-  }
-
-  private handlePending(
-    id: string,
-    stats: Stats
-  ): void {
-    this.nodes.updatePending(
-      id, stats,
-      (err: Error | string, pending: Pending) => {
-        if (err) {
-          console.error('API', 'TXS', 'Pending error:', err)
-        }
-
-        if (pending) {
-          this.clientWrite({
-            action: 'pending',
-            data: pending
-          })
-
-          console.success(
-            'API', 'TXS', 'Pending:',
-            pending['pending'],
-            'from:', pending.id
-          )
-        }
-      }
-    )
-  }
-
-  private handleValidators(
-    validators: Validators
-  ): void {
-    if (validators && validators.registered) {
-      validators.registered.forEach(validator => {
-        validator.registered = true
-
-        // trust registered validators and signers - not safe
-        if (
-          validator.address &&
-          trusted.indexOf(validator.address) === -1
-        ) {
-          trusted.push(validator.address)
-        }
-
-        if (
-          validator.signer &&
-          trusted.indexOf(validator.signer) === -1
-        ) {
-          trusted.push(validator.signer)
-        }
-
-        const search = {id: validator.address}
-        const index: number = this.nodes.getIndex(search)
-        const node: Node = this.nodes.getNodeOrNew(search, validator)
-
-        if (index < 0) {
-          // only if new node
-          node.integrateValidatorData(validator)
-        }
-
-        node.setValidatorData(validator)
-
-        if (validators.elected.indexOf(validator.address) > -1) {
-          node.setValidatorElected(true)
-        }
-
-        node.setValidatorRegistered(true)
-      })
-    }
-  }
-
-  private handleStatsUpdate(
-    id: string,
-    stats: Stats
-  ): void {
-    this.nodes.updateStats(
-      id, stats,
-      (err: Error | string, basicStats: BasicStatsResponse) => {
-        if (err) {
-          console.error('API', 'STA', 'Stats error:', err)
-        } else {
-
-          if (basicStats) {
-            this.clientWrite({
-              action: 'stats',
-              data: basicStats
-            })
-
-            console.info(
-              'API', 'STA',
-              'Stats from:', id
-            )
-          }
-        }
-      })
-  }
-
-  private handleLatency(
-    id: string,
-    latency: number
-  ): void {
-    this.nodes.updateLatency(
-      id, latency,
-      (err: Error | string, latency: Latency): void => {
-        if (err) {
-          console.error('API', 'PIN', 'Latency error:', err)
-        }
-
-        if (latency) {
-          console.info(
-            'API', 'PIN',
-            'Latency:', latency.latency,
-            'from:', id
-          )
-        }
-      }
-    )
-  }
-
-  private handleNodeEnd(
-    id: string,
-    ip: string
-  ): void {
-
-    // it this was caused by a failed auth do not try this
-    if (this.danglingConnections[id]) {
-      delete (this.danglingConnections[id])
-    } else {
-      this.nodes.inactive(
-        id,
-        (err: Error | string, nodeStats: NodeStats
-        ) => {
-          if (err) {
-            console.error(
-              'API', 'CON',
-              'Connection with:', id,
-              'ended.', 'Error:', err
-            )
-          } else {
-            this.clientWrite({
-              action: 'inactive',
-              data: nodeStats
-            })
-
-            console.success(
-              'API', 'CON', 'Node:',
-              `'${nodeStats.name}'`, `(${id})`,
-              'disconnected.'
-            )
-          }
-        })
-    }
-
-    console.success(
-      'API', 'CON', 'Node Close:',
-      ip, `'${id}'`
-    )
-  }
-
-  private handleNodePing(
-    id: string,
-    stats: NodePing,
-    spark: Primus.spark
-  ): void {
-    const start = (!_.isUndefined(stats.clientTime) ? stats.clientTime : null)
-
-    spark.emit(
-      'node-pong',
-      <NodePong>{
-        clientTime: start,
-        serverTime: _.now()
-      }
-    )
-
-    this.statistics.add(Sides.Node, Directions.Out)
-
-    console.info('API', 'PIN', 'Ping from:', id)
-  }
-
-  private handleAuth(
-    id: string,
-    proof: Proof,
-    stats: InfoWrapped,
-    spark: Primus.spark
-  ): boolean {
-
-    if (
-      banned.indexOf(spark.address.ip) >= 0 ||
-      !Server.isAuthorize(proof, stats)
-    ) {
-      console.error(
-        'API', 'CON', 'Node Closed: wrong auth',
-        `'${stats.id}'`, `(${spark.id})`,
-        'address:', proof.address
-      )
-
-      spark.end(undefined, {reconnect: false})
-
-      return false
-    }
-
-    delete (this.danglingConnections[id])
-
-    console.info(
-      'API', 'CON',
-      'Hello', stats.id
-    )
-
-    return true
-  }
-
-  private handleNode(
-    id: string,
-    proof: Proof,
-    stats: InfoWrapped,
-    spark: Primus.spark
-  ): void {
-
-    const nodeData: NodeData = {
-      id,
-      address: proof.address,
-      ip: spark.address.ip,
-      spark: spark.id,
-      latency: spark.latency || 0
-    }
-
-    this.nodes.add(
-      <NodeInformation>{
-        nodeData,
-        stats
-      },
-      (err: Error | string, info: NodeInfo): void => {
-        if (err) {
-          console.error('API', 'CON', 'Connection error:', err)
-          return
-        }
-
-        if (info) {
-          spark.emit('ready')
-
-          this.statistics.add(Sides.Node, Directions.Out)
-
-          console.success(
-            'API', 'CON', 'Node',
-            `'${stats.id}'`, `(${spark.id})`, 'Connected')
-
-          this.clientWrite({
-            action: 'add',
-            data: info
-          })
-        }
-      })
-  }
-
-  private handleReady(
-    id: string,
-    spark: Primus.spark
-  ): void {
-    spark.emit(
-      'init',
-      {nodes: this.nodes.all()}
-    )
-
-    this.statistics.add(Sides.Client, Directions.Out)
-
-    this.nodes.getCharts()
-
-    console.success(
-      'API', 'CON',
-      'Client', `'${id}'`,
-      'Connected'
-    )
-  }
-
-  private handleClientPong(
-    data: ClientPong,
-    spark: Primus.spark
-  ): void {
-    const serverTime = _.get(data, 'serverTime', 0)
-    const latency = Math.ceil((_.now() - serverTime) / 2)
-
-    spark.emit(
-      'client-latency',
-      {latency: latency}
-    )
-
-    this.statistics.add(Sides.Client, Directions.Out)
-  }
-
-  private handleClientEnd(
-    id: string,
-    ip: string
-  ): void {
-    console.success(
-      'API', 'CON', 'Client Close:',
-      ip, `'${id}'`
-    )
+    this.controller.init()
   }
 }
