+++ src/server/utils/isInputValid.ts
@@ -1,0 +1,10 @@
+import { Wrapper } from "../interfaces/Wrapper";
+import _ from "lodash";
+
+export function isInputValid(
+  stats: Wrapper
+): boolean {
+  return (
+    !_.isUndefined(stats) && !_.isUndefined(stats.id)
+  )
+}
